+++ ltr/aggregators/vote_aggregator.h
@@ -21,7 +21,7 @@ class VoteAggregator : public Aggregator {
   double aggregate(const vector<double>& labels,
                    const vector<double>& weights) {
     map<double, double> voting;
-    for (int label_index = 0; label_index < labels.size(); ++label_index) {
+    for (int label_index = 0; label_index < (int)labels.size(); ++label_index) {
       voting[labels[label_index]] += weights[label_index];
     }
     double result = voting.begin()->first;
