+++ ltr/scorers/composition_scorers/order_statistic_composition_scorer.cc
@@ -26,7 +26,7 @@ namespace ltr {
     // if weight[i] < 0 then take it 0
     double sum_weight = 0.0;
     vector<pair<double, double> > weighted_scores(size());
-    for (size_t i = 0; i < size(); ++i) {
+    for (int i = 0; i < size(); ++i) {
       sum_weight += NonNegative(at(i).weight);
       weighted_scores[i] = make_pair(((*at(i).scorer)(object)),
         NonNegative(at(i).weight));
